Sudden Break - Disaster Management System( College Project)
The sudden break is a web and android based disaster management system. It consists of Smartphone application and a website. The project does not avert or eliminate the threats; instead it focuses on creating plans to decrease disasters. The project consists of various departments and users through which accurate information can be communicated easily. The project is done using python, (java) android as front end and My SQL as back end.